+++ common/src/main/java/me/giiena/config/impl/client/screen/component/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package me.giiena.config.impl.client.screen.component;
+
+import org.jetbrains.annotations.ApiStatus;