+++ common/src/main/java/me/giiena/config/impl/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package me.giiena.config.impl;
+
+import org.jetbrains.annotations.ApiStatus;