+++ common/src/main/java/me/giiena/config/impl/platform/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package me.giiena.config.impl.platform;
+
+import org.jetbrains.annotations.ApiStatus;