+++ fabric/src/main/java/me/giiena/config/impl/integration/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package me.giiena.config.impl.integration;
+
+import org.jetbrains.annotations.ApiStatus;